+++ docs/source/changelog.md
@@ -6,8 +6,8 @@
 ### ✨ Features
 - Update the wakepy CLI printout: Adds the used Method and activated Mode to the printout ([#434](https://github.com/fohrloop/wakepy/pull/434))
 - Added `-v` (INFO) and `-vv` (DEBUG) verbosity flags for the [wakepy CLI command](https://wakepy.readthedocs.io/stable/cli-api.html). ([#439](https://github.com/fohrloop/wakepy/pull/439))
-- Added [MethodInfo](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodInfo) objects, which are now used in [Mode.active_method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.Mode.active_method), [Mode.used_method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.Mode.used_method),  [ActivationResult.active_method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.ActivationResult.active_method) instead of them being strings.  ([#459](https://github.com/fohrloop/wakepy/pull/459), [#460](https://github.com/fohrloop/wakepy/pull/460))
-- The [MethodActivationResult](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodActivationResult) has now a [method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodActivationResult.method) attribute which is also a [MethodInfo](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodInfo) instance. ([#460](https://github.com/fohrloop/wakepy/pull/460))
+- Add [Mode.method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.Mode.method), [ActivationResult.method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.ActivationResult.method) and [MethodActivationResult.method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodActivationResult.method) attributes, which are instances of [MethodInfo](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodInfo) ([#459](https://github.com/fohrloop/wakepy/pull/459), [#460](https://github.com/fohrloop/wakepy/pull/460), [#464](https://github.com/fohrloop/wakepy/pull/464))
+- Make [Mode.active_method](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.Mode.active_method) a [MethodInfo](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.MethodInfo) instance (was a string) ([#459](https://github.com/fohrloop/wakepy/pull/459))
 
 ### ✨ Minor Enhancements
 - Better error messages: When selected Method is not part of the selected Mode ([#427](https://github.com/fohrloop/wakepy/pull/427)) and when a D-Bus -based method fails ([#438](https://github.com/fohrloop/wakepy/pull/438))
@@ -21,6 +21,9 @@
 - Simplify and clarify the documentation of the [Mode](https://wakepy.readthedocs.io/stable/api-reference.html#wakepy.Mode) ([#458](https://github.com/fohrloop/wakepy/pull/458))
 
 ### 🚨 Backwards incompatible changes
+- Deprecate `Mode.used_method`. Use `Mode.method`, instead ([464](https://github.com/fohrloop/wakepy/pull/464))
+- Deprecate `Mode.activation_result`. Use `Mode.result`, instead ([464](https://github.com/fohrloop/wakepy/pull/464))
+- Deprecate `ActivationResult.active_method`. Use `ActivationResult.method`, instead ([464](https://github.com/fohrloop/wakepy/pull/464))
 - Make the `Mode._from_name()` and `Mode._method_classes` private; Not part of the public API anymore ([#458](https://github.com/fohrloop/wakepy/pull/458))
 - The `Mode.active_method`, `Mode.used_method`, are now instances of the new ModeInfo (previously strings) ([#460](https://github.com/fohrloop/wakepy/pull/460))
 - The `ActivationResult.mode_name` is now always a string (instead of being a ModeName) ([#462](https://github.com/fohrloop/wakepy/pull/462))
@@ -30,7 +33,6 @@
 - Make Run Tox fail on Windows pipelines if pytest fails ([#442](https://github.com/fohrloop/wakepy/pull/442)). There was a bug in the workflow file which could make the tests pass on Windows even if pytest fails.
 - Fix flaky tests on pypy+windows ([#447](https://github.com/fohrloop/wakepy/pull/447))
 - Remove the AttributeError traceback when building docs on Windows/MacOS ([#449](https://github.com/fohrloop/wakepy/pull/449))
-
 
 
 ## wakepy 0.10.2